Add new module to store all logic, update readme file
